Add reference scores
