+++ .gitignore
@@ -9,3 +9,4 @@ venv/
 __init__.pyc
 coverage.xml
 docs/api/
+tests/assests/linkability/anonymization_metrics/
