+++ setup.cfg
@@ -41,6 +41,7 @@ addopts =
     --cov-fail-under=100
     --cov-report xml
     --cov-report term-missing
+    --ignore tests/assests/linkability/
 xfail_strict = true
 
 [flake8]
